Hel 669 670 671/changement design popup gestion listes (#735)
* Revue du design de la popup

* Alignement de la popup à droite du bouton d’action

* Extraction des fonctionnalités de la popup des favoris dans un fichier dédié
