+++ src/frontend/configuration/wording/Wording.tsx
@@ -446,7 +446,8 @@ export interface Wording {
   readonly ETOILE_ETAB_DANS_LISTE: string;
   readonly ETOILE_ETAB_PAS_DANS_LISTE: string;
   readonly ETOILE_MES_LISTES: string;
-  readonly ETOILE_NOUVELLE_LISTE: string;
+  readonly ETOILE_NOUVELLE_LISTE_LABEL: string;
+  readonly ETOILE_NOUVELLE_LISTE_BOUTON: string;
   readonly ETOILE_MAX_LISTE_ATTEINT: string;
   readonly ETOILE_ERREUR_MODIF_ETAB: string;
